Upgrade Scala.js to 1.15.0, allow Reader tests now
