+++ sconfig/shared/src/test/scala/org/ekrich/config/impl/ConfigFactoryTest.scala
@@ -1,25 +1,27 @@
 package org.ekrich.config.impl
 
-// import java.io.StringReader
+import java.io.StringReader
 
 import org.junit.Assert._
 import org.junit.Test
 
+import org.ekrich.config.ConfigException
 import org.ekrich.config.ConfigFactory
+import org.ekrich.config.ConfigParseOptions
 
 class ConfigFactoryTest {
+  // use string
+  val fileStr =
+    """
+      |# test01.properties file
+      |fromProps.abc=abc
+      |fromProps.one=1
+      |fromProps.bool=true
+      |fromProps.specialChars=hello^^
+      """.stripMargin
 
-  // uncomment when Scala.js has Reader
-  // val filename = "/test01.properties"
-  // val fileStr =
-  //   """
-  //     |# test01.properties file
-  //     |fromProps.abc=abc
-  //     |fromProps.one=1
-  //     |fromProps.bool=true
-  //     |fromProps.specialChars=hello^^
-  //     """.stripMargin
-  // var test01Reader = new StringReader(fileStr)
+  // create Reader
+  var test01Reader = new StringReader(fileStr)
 
   @Test def parseString: Unit = {
     val configStr =
@@ -59,25 +61,25 @@ class ConfigFactoryTest {
     assertEquals(pattern, "default")
   }
 
-  // @Test
-  // def parse(): Unit = {
-  //   val filename = "/test01.properties"
-  //   val config = ConfigFactory.parseReader(
-  //     test01Reader,
-  //     ConfigParseOptions.defaults
-  //       .setSyntaxFromFilename(filename)
-  //   )
-  //   assertEquals("hello^^", config.getString("fromProps.specialChars"))
-  // }
+  @Test
+  def parse(): Unit = {
+    val filename = "/test01.properties"
+    val config = ConfigFactory.parseReader(
+      test01Reader,
+      ConfigParseOptions.defaults
+        .setSyntaxFromFilename(filename)
+    )
+    assertEquals("hello^^", config.getString("fromProps.specialChars"))
+  }
 
-  // @Test
-  // def parseIncorrectFormat(): Unit = {
-  //   val e = assertThrows(
-  //     classOf[ConfigException.Parse],
-  //     () => ConfigFactory.parseReader(test01Reader)
-  //   )
-  //   assertTrue(
-  //     e.getMessage.contains("Expecting end of input or a comma, got '^'")
-  //   )
-  // }
+  @Test
+  def parseIncorrectFormat(): Unit = {
+    val e = assertThrows(
+      classOf[ConfigException.Parse],
+      () => ConfigFactory.parseReader(test01Reader)
+    )
+    assertTrue(
+      e.getMessage.contains("Expecting end of input or a comma, got '^'")
+    )
+  }
 }
